requirements.txt
streamlit
google-generativeai
chromadb
pypdf2
langchain-text-splitters
pydantic
python-dotenv